Avoid tests killing entire process groups

See merge request gnutls/gnutls!1463

+++ tests/key-material-dtls.c
@@ -83,8 +83,10 @@ static void terminate(void)
 {
 	int status = 0;
 
-	kill(child, SIGTERM);
-	wait(&status);
+	if (child) {
+		kill(child, SIGTERM);
+		wait(&status);
+	}
 	exit(1);
 }
 
@@ -141,7 +143,7 @@ static void client(int fd)
 
 	if (ret < 0) {
 		fail("client: Handshake failed: %s\n", strerror(ret));
-		terminate();
+		exit(1);
 	} else {
 		if (debug)
 			success("client: Handshake was completed\n");
@@ -155,13 +157,13 @@ static void client(int fd)
 	ret = gnutls_cipher_get(session);
 	if (ret != GNUTLS_CIPHER_AES_128_CBC) {
 		fprintf(stderr, "negotiated unexpected cipher: %s\n", gnutls_cipher_get_name(ret));
-		terminate();
+		exit(1);
 	}
 
 	ret = gnutls_mac_get(session);
 	if (ret != GNUTLS_MAC_SHA1) {
 		fprintf(stderr, "negotiated unexpected mac: %s\n", gnutls_mac_get_name(ret));
-		terminate();
+		exit(1);
 	}
 
 	iv_size = 16;
@@ -174,7 +176,7 @@ static void client(int fd)
 	if (ret < 0) {
 		fprintf(stderr, "error in %d\n", __LINE__);
 		gnutls_perror(ret);
-		terminate();
+		exit(1);
 	}
 	p = key_material;
 
@@ -183,33 +185,33 @@ static void client(int fd)
 	if (ret < 0) {
 		fprintf(stderr, "error in %d\n", __LINE__);
 		gnutls_perror(ret);
-		terminate();
+		exit(1);
 	}
 
 	if (memcmp(wseq_number, "\x00\x01\x00\x00\x00\x00\x00\x01", 8) != 0) {
 		dump("wseq:", wseq_number, 8);
 		fprintf(stderr, "error in %d\n", __LINE__);
-		terminate();
+		exit(1);
 	}
 
 	ret = gnutls_record_get_state(session, 1, &read_mac_key, &read_iv, &read_cipher_key, rseq_number);
 	if (ret < 0) {
 		fprintf(stderr, "error in %d\n", __LINE__);
 		gnutls_perror(ret);
-		terminate();
+		exit(1);
 	}
 
 	if (memcmp(rseq_number, "\x00\x01\x00\x00\x00\x00\x00\x01", 8) != 0) {
 		dump("rseq:", rseq_number, 8);
 		fprintf(stderr, "error in %d\n", __LINE__);
-		terminate();
+		exit(1);
 	}
 
 	if (hash_size != mac_key.size || memcmp(p, mac_key.data, hash_size) != 0) {
 		dump("MAC:", mac_key.data, mac_key.size);
 		dump("Block:", key_material, block_size);
 		fprintf(stderr, "error in %d\n", __LINE__);
-		terminate();
+		exit(1);
 	}
 	p+= hash_size;
 
@@ -217,19 +219,19 @@ static void client(int fd)
 		dump("MAC:", read_mac_key.data, read_mac_key.size);
 		dump("Block:", key_material, block_size);
 		fprintf(stderr, "error in %d\n", __LINE__);
-		terminate();
+		exit(1);
 	}
 	p+= hash_size;
 
 	if (key_size != cipher_key.size || memcmp(p, cipher_key.data, key_size) != 0) {
 		fprintf(stderr, "error in %d\n", __LINE__);
-		terminate();
+		exit(1);
 	}
 	p+= key_size;
 
 	if (key_size != read_cipher_key.size || memcmp(p, read_cipher_key.data, key_size) != 0) {
 		fprintf(stderr, "error in %d\n", __LINE__);
-		terminate();
+		exit(1);
 	}
 	p+= key_size;
 
@@ -246,13 +248,13 @@ static void client(int fd)
 	if (ret < 0) {
 		fprintf(stderr, "error in %d\n", __LINE__);
 		gnutls_perror(ret);
-		terminate();
+		exit(1);
 	}
 
 	if (memcmp(wseq_number, "\x00\x01\x00\x00\x00\x00\x00\x06", 8) != 0) {
 		dump("wseq:", wseq_number, 8);
 		fprintf(stderr, "error in %d\n", __LINE__);
-		terminate();
+		exit(1);
 	}
 
 	memset(rseq_number, 0xAA, sizeof(rseq_number));
@@ -260,13 +262,13 @@ static void client(int fd)
 	if (ret < 0) {
 		fprintf(stderr, "error in %d\n", __LINE__);
 		gnutls_perror(ret);
-		terminate();
+		exit(1);
 	}
 
 	if (memcmp(rseq_number, "\x00\x01\x00\x00\x00\x00\x00\x01", 8) != 0) {
 		dump("rseq:", rseq_number, 8);
 		fprintf(stderr, "error in %d\n", __LINE__);
-		terminate();
+		exit(1);
 	}
 	gnutls_bye(session, GNUTLS_SHUT_WR);
